refactor WorkspaceController by moving mapper functions into separate class
