+++ api/src/test/java/org/pmiops/workbench/api/ConceptSetsControllerTest.java
@@ -35,8 +35,9 @@ import org.pmiops.workbench.db.dao.UserDao;
 import org.pmiops.workbench.db.dao.UserRecentResourceService;
 import org.pmiops.workbench.db.dao.UserService;
 import org.pmiops.workbench.db.dao.WorkspaceDao;
-import org.pmiops.workbench.db.dao.WorkspaceService;
-import org.pmiops.workbench.db.dao.WorkspaceServiceImpl;
+import org.pmiops.workbench.workspaces.WorkspaceMapper;
+import org.pmiops.workbench.workspaces.WorkspaceService;
+import org.pmiops.workbench.workspaces.WorkspaceServiceImpl;
 import org.pmiops.workbench.db.model.CdrVersion;
 import org.pmiops.workbench.db.model.User;
 import org.pmiops.workbench.exceptions.BadRequestException;
@@ -146,6 +147,9 @@ public class ConceptSetsControllerTest {
   WorkspaceService workspaceService;
 
   @Autowired
+  WorkspaceMapper workspaceMapper;
+
+  @Autowired
   ConceptSetDao conceptSetDao;
 
   @Autowired
@@ -188,7 +192,7 @@ public class ConceptSetsControllerTest {
 
 
   @TestConfiguration
-  @Import({WorkspaceServiceImpl.class, CohortCloningService.class, CohortFactoryImpl.class,
+  @Import({WorkspaceServiceImpl.class, WorkspaceMapper.class, CohortCloningService.class, CohortFactoryImpl.class,
       UserService.class, ConceptSetsController.class, WorkspacesController.class, ConceptSetService.class})
   @MockBean({ConceptBigQueryService.class, FireCloudService.class, CloudStorageService.class,
       ConceptSetService.class, UserRecentResourceService.class, ComplianceService.class})
@@ -210,7 +214,7 @@ public class ConceptSetsControllerTest {
     conceptSetsController = new ConceptSetsController(workspaceService, conceptSetDao, conceptDao,
         conceptBigQueryService, userRecentResourceService, userProvider, CLOCK);
     WorkspacesController workspacesController =
-        new WorkspacesController(workspaceService, cdrVersionDao, cohortDao, cohortFactory, conceptSetDao,
+        new WorkspacesController(workspaceService, workspaceMapper, cdrVersionDao, cohortDao, cohortFactory, conceptSetDao,
                 userDao, userProvider, fireCloudService, cloudStorageService, CLOCK, userService,
                 userRecentResourceService);
